requirements.txt
# Core
anthropic>=0.40.0
python-dotenv>=1.0.0
aiohttp>=3.9.0

# Data Ingestion
feedparser>=6.0.0
tavily-python>=0.5.0

# Stock Data
yfinance>=0.2.30

# Storage — Local
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Storage — GCP (BigQuery + Vertex AI)
google-cloud-bigquery>=3.25.0
google-cloud-bigquery-storage>=2.25.0
google-cloud-aiplatform>=1.60.0
google-cloud-secret-manager>=2.20.0
google-auth>=2.30.0

# Processing
simhash>=2.1.2

# Frontend
streamlit>=1.40.0
plotly>=5.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
